+++ components/Buttons/CopyButton.tsx
@@ -1,4 +1,5 @@
 import { Box, IconButton, Tooltip } from '@chakra-ui/react';
+import { useState } from 'react';
 import { FiCopy } from 'react-icons/fi';
 
 interface Props {
