+++ components/Code/DisplayCode.tsx
@@ -1,11 +1,10 @@
 import { Box } from '@chakra-ui/react';
-import Highlight, { Language, defaultProps } from 'prism-react-renderer';
+import Highlight, { defaultProps } from 'prism-react-renderer';
 import nightOwl from 'prism-react-renderer/themes/nightOwl';
 
 import { ILanguage, PasteType } from 'types';
 
-import ActionsButton from './ActionsButton';
-import CopyButton from './CopyButton';
+import ActionsButton from '../Buttons/ActionsButton';
 
 interface Props {
   paste: PasteType;
