+++ pages/_document.tsx
@@ -1,4 +1,3 @@
-// pages/_document.js
 import NextDocument, { Head, Html, Main, NextScript } from 'next/document';
 
 import { ColorModeScript } from '@chakra-ui/react';
