+++ pages/user/pastes/[userId].tsx
@@ -25,14 +25,16 @@ import {
 } from 'react-icons/hi';
 import useLocalStorage from 'use-local-storage';
 
-import NoPastes from 'components/CodePastes/NoPastes';
 import Paste from 'components/CodePastes/Paste';
 import Layout from 'components/Layout';
+import NoPastes from 'components/Pastes/NoPastes';
 
 import reduceTitleLength from 'utils/reduceTitleLength';
 import supabaseClient from 'utils/supabase';
 
 import { PasteType, User } from 'types';
+
+components / Pastes / Paste;
 
 const links = [
   {
